cdk/examples/cdk-example-s3/requirements.txt
-e .
pytest
aws-cdk.core
aws-cdk.aws-s3